Improvements to bitmap caching and corrections for AGG bitmap drawing

+++ IPlug/IGraphics/IGraphics.cpp
@@ -117,7 +117,7 @@ void IGraphics::SetFromStringAfterPrompt(IControl* pControl, IParam* pParam, con
   }
 }
 
-void IGraphics::AttachBackground(const char* name, double scale)
+void IGraphics::AttachBackground(const char* name, int scale)
 {
   IBitmap bg = LoadBitmap(name, 1, false, scale);
   mControls.Insert(0, new IBitmapControl(mPlug, 0, 0, -1, bg, IBlend::kBlendClobber));
